move MessageContainer to shared package
